+++ Repositories/IPriceRepository.cs
@@ -5,5 +5,6 @@ namespace AthenaFinance.Api.Repositories;
 public interface IPriceRepository
 {
     Task<IList<EodPrice>> GetPricesAsync(int securityId, DateOnly from, DateOnly to);
+    Task<EodPrice?> GetLatestAsync(int securityId);
     Task UpsertAsync(int securityId, IEnumerable<EodPrice> prices);
 }